MOB-987 make resetUser graceful when dismissing announcements and inline delete steps for detox
